Move function to postgresql.py file
