+++ src/hdx/database/postgresql.py
@@ -1,9 +1,11 @@
 """PostgreSQL specific utilities"""
 
 import logging
+import subprocess
 import time
+from os import environ
 
-from .dburi import remove_driver_from_uri
+from .dburi import get_params_from_connection_uri, remove_driver_from_uri
 
 try:
     import psycopg
@@ -17,6 +19,10 @@ except ImportError:
     )
 
 logger = logging.getLogger(__name__)
+
+
+class PostgresError(Exception):
+    pass
 
 
 def wait_for_postgresql(db_uri: str) -> None:
@@ -39,3 +45,50 @@ def wait_for_postgresql(db_uri: str) -> None:
             break
         except psycopg.OperationalError:
             time.sleep(1)
+
+
+def restore_from_pgfile(db_uri: str, pg_restore_file: str) -> str:
+    """Restore database from a pg_restore file created by pg_backup.
+
+    Args:
+        db_uri (str): Connection URI.
+        pg_restore_file (str): Path to the pg_restore database file
+
+    Returns:
+        None
+    """
+    db_params = get_params_from_connection_uri(db_uri)
+    subprocess_params = ["pg_restore", "-c"]
+    for key, value in db_params.items():
+        match key:
+            case "database":
+                subprocess_params.append("-d")
+            case "host":
+                subprocess_params.append("-h")
+            case "port":
+                subprocess_params.append("-p")
+            case "username":
+                subprocess_params.append("-U")
+            case _:
+                continue
+        subprocess_params.append(f"{value}")
+
+    subprocess_params.append(f"{pg_restore_file}")
+    env = environ.copy()
+    password = db_params.get("password")
+    if password:
+        env["PGPASSWORD"] = password
+    process = subprocess.run(
+        subprocess_params, env=env, capture_output=True, encoding="utf-8"
+    )
+    try:
+        process.check_returncode()
+    except subprocess.CalledProcessError as ex:
+        command = " ".join(subprocess_params)
+        raise PostgresError(
+            f"{command} failed. Return code: {process.returncode}"
+        ) from ex
+
+    for line in process.stdout.splitlines():
+        logger.info(line)
+    return process.stdout
